Refactor common code to fileDownload.js
